pyproject.toml
[tool.poetry]
name = "support_test"
version = "0.1.0"
description = ""
authors = ["securitsa <khitricov.e@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.2"
djangorestframework = "^3.13.1"
psycopg2-binary = "^2.9.3"
transaction = "^3.0.1"
python-dotenv = "^0.19.2"
pre-commit = "^2.17.0"
djangorestframework-simplejwt = "^5.0.0"
celery = {extras = ["redis"], version = "^5.2.5"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
